feat: employee read, update, delete, dashboard

+++ backend/src/graphql/resolvers/employee.resolver.js
@@ -1,16 +1,25 @@
 const Employee = require('../../model/employee')
+const { isValidObjectId } = require('mongoose')
 
 const resolvers = {
     Query: {
          // EMPLOYEE
         getAllEmployees: async () => {
             const employees = await Employee.find();
-            return employees;
+            if (!employees || employees.length === 0)
+                return {employees: employees, success: false, message: "No employees found"};
+            return {employees: employees, success: true, message: "All employees fetched successfully"};
         },
 
         searchEmployeeById: async (_, {id}) => {
-            const employee = await Employee.findOne({ id: id })
-            return employee;
+            if(!id || !isValidObjectId(id))
+                return {employee: null, success: false, message: "Invalid or missing employee ID"};
+
+            const employee = await Employee.findOne({ _id: id })
+            if(!employee)
+                return {employee: null, success: false, message: "Employee not found"};
+
+            return {employee: employee, success: true, message: "Employee fetched successfully"};
         },
 
         searchEmployeeByDesignationOrDepartment: async (_, {designation, department}) => {
@@ -22,7 +31,9 @@ const resolvers = {
             if(department) condition.department = department; 
 
             const employees = await Employee.find(condition)
-            return employees;
+            if (!employees || employees.length === 0)
+                return {employees: employees, success: false, message: "No matching employees found"};
+            return {employees: employees, success: true, message: "Searched fetched successfully"};
         }
     },
 
@@ -32,6 +43,9 @@ const resolvers = {
             first_name, last_name, email, gender, designation, 
             salary, date_of_joining, department, employee_photo
         }) => {
+            const existingEmployee = await Employee.findOne({ email: email });
+            if (existingEmployee) return { success: false, message: "Email has already been taken", employee: null };
+
             const newEmployee = new Employee({
                 first_name: first_name,
                 last_name: last_name,
@@ -45,13 +59,19 @@ const resolvers = {
             });
 
             await newEmployee.save();
-            return newEmployee;
+            return { success: true, message: "Employee created successfully", employee: newEmployee };
         },
 
         updateEmployee: async (_, {
             id, first_name, last_name, email, gender, designation, 
             department, salary, date_of_joining, employee_photo
         }) => {
+            if(!id || !isValidObjectId(id))
+                return {employee: null, success: false, message: "Invalid or missing employee ID"};
+
+            const existingEmployee = await Employee.findOne({ email: email });
+            if (existingEmployee) return { success: false, message: "Email has already been taken", employee: null };
+
             const updatedEmployee = await Employee.findOneAndUpdate(
                 {_id: id},  
                 // $set indicates which fields to update
@@ -73,15 +93,29 @@ const resolvers = {
                     runValidators: true, // run mongoose validation
                     new: true // return updated object
                 } 
-    
             );
-            return updatedEmployee;
+
+            if(!updatedEmployee) return { success: false, message: "Employee not found", employee: null };
+            return { 
+                success: true, 
+                message: `Employee updated successfully`,
+                employee: updatedEmployee
+            };
         },
 
         deleteEmployee: async (_, {id}) => {
+            if(!id || !isValidObjectId(id))
+                return {employee: null, success: false, message: "Invalid or missing employee ID"};
+            
             const deletedEmployee = await Employee.findOneAndDelete({_id: id})
             // findOneAndDelete returns doc if deleted, returns null if not found
-            return deletedEmployee != null // success = return true, fail = return false
+
+            if(!deletedEmployee) return { success: false, message: "Employee not found", employee: null };
+            return { 
+                success: true, 
+                message: `Employee deleted successfully`, 
+                employee: deletedEmployee 
+            }
         }
     }
